- Authentication with tokens - Local dev settings update - Recordings GET pagination to minimize request sizes - Refactoring and removing dead code
