+++ restapi/views/__init__.py
@@ -1,0 +1,2 @@
+from .site import *
+from .model import *
