+++ tarteel/urls.py
@@ -7,6 +7,7 @@ from django.contrib import admin
 from django.urls import path, include
 # REST
 from rest_framework import routers
+from rest_framework.authtoken import views as authviews
 # Tarteel
 import audio.views
 import evaluation.views
@@ -19,17 +20,10 @@ router = routers.DefaultRouter()
 
 urlpatterns = [
     # Rest API v1
-    path('api/v1/', include('restapi.urls')),
+    path('', include('restapi.urls')),
     # Top Level API
     url(r'^admin/', admin.site.urls),
     url(r'^api-auth/', include('rest_framework.urls', namespace='rest_framework')),
-    url(r'^api/recordings/', restapi.views.AnnotatedRecordingList.as_view(),
-        name='file-upload'),
-    url(r'^api/get_ayah/', restapi.views.GetAyah.as_view(), name='get_ayah'),
-    url(r'^api/index/', restapi.views.Index.as_view(), name='api_index'),
-    url(r'^api/about/', restapi.views.About.as_view(), name='api_about'),
-    url(r'^api/surah/(?:(?P<num>\d+)/)?$', restapi.views.GetSurah.as_view(), name='get_Surah'),
-    url(r'^api/profile/(?P<session_key>[\w-]+)/', restapi.views.Profile.as_view(), name='profile_api'),
     url(r'^api/demographics/', restapi.views.DemographicInformationViewList.as_view(),
         name='demographic'),
     url(r'^api/evaluator/', evaluation.views.EvaluationList.as_view(), name="evaluation"),
@@ -37,17 +31,10 @@ urlpatterns = [
     url(r'^api/v2/submit_evaluation', restapi.views.EvaluationSubmission.as_view(), name="v2_evaluation_submission"),
     url(r'^api/get_evaluations_count/', evaluation.views.get_evaluations_count,
         name="get_evaluations_count"),
-    url(r'^get_ayah/', audio.views.get_ayah),
     url(r'^get_ayah_translit/', audio.views.get_ayah_translit),
     url(r'^get_total_count/', restapi.views.RecordingsCount.as_view(),
         name='recordingscount'),
-    url(r'^download-audio/', audio.views.download_audio),
     url(r'^api/download-audio/', restapi.views.DownloadAudio.as_view()),
-    url(r'^sample-recordings/', audio.views.sample_recordings),
-    url(r'^download-full-dataset-csv/', audio.views.download_full_dataset_csv),
-    # Audio App (Main Page)
-    url(r'^about/', audio.views.about),
-    url(r'^profile/(?P<session_key>[\w-]+)/', audio.views.profile),
     # Evaluation tools
     url(r'^evaluator/', evaluation.views.evaluator),
     url(r'^evaluation/evaluator/', evaluation.views.evaluator),
@@ -56,6 +43,7 @@ urlpatterns = [
         name='tajweed-evaluation'),
     # Django-allauth Login
     url(r'^accounts/', include('allauth.urls')),
+    url(r'^api-token-auth/', authviews.obtain_auth_token),
 ]
 
 urlpatterns += router.urls
